server/requirements.txt
Django==1.6.1
Pillow==2.2.2
Unidecode==0.04.14
djangorestframework==2.3.10
python-social-auth
wsgiref==0.1.2
